% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ubestarfm-package}
\alias{ubestarfm}
\alias{ubestarfm-package}
\title{ubestarfm package}
\description{
Trains reusable reference-pair models and predicts fine-resolution land surface temperature using the unbiased ESTARFM algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yuyi13/ubESTARFM}
  \item Report bugs at \url{https://github.com/yuyi13/ubESTARFM/issues}
}

}
\author{
\strong{Maintainer}: Yi Yu \email{yi.yu1@anu.edu.au}

}
\keyword{internal}
