% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ubestarfm_load_model}
\alias{ubestarfm_load_model}
\title{Load a reusable ubESTARFM model}
\usage{
ubestarfm_load_model(path)
}
\arguments{
\item{path}{An RDS model path.}
}
\value{
A validated \code{ubestarfm_model}.
}
\description{
Load a reusable ubESTARFM model
}
