% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ubestarfm_read_raster}
\alias{ubestarfm_read_raster}
\title{Read one raster layer}
\usage{
ubestarfm_read_raster(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{A file path or a \code{terra::SpatRaster}.}

\item{name}{Name used in validation messages.}
}
\value{
A list containing the source raster and a row-column matrix.
}
\description{
Read one raster layer
}
