% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ubestarfm_validate_geometry}
\alias{ubestarfm_validate_geometry}
\title{Validate raster geometry}
\usage{
ubestarfm_validate_geometry(rasters)
}
\arguments{
\item{rasters}{A list of \code{terra::SpatRaster} objects.}
}
\value{
\code{TRUE}, invisibly, when all geometries match.
}
\description{
Validate raster geometry
}
