% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ubESTARFM-package}
\alias{ubESTARFM}
\alias{ubESTARFM-package}
\title{ubESTARFM package}
\description{
Trains reusable reference-pair models and predicts fine-resolution land surface temperature using the unbiased ESTARFM algorithm.
}
\author{
\strong{Maintainer}: Yi Yu \email{yi.yu1@anu.edu.au}

}
\keyword{internal}
